Implement basic interactivity
I have added some methods to the GameBoard class that collect input from the player and use that
to place the desired symbol. First the game asks for a row, and passes it to the row_select method.
Next, it asks for a column and passes it to the column select method. Finally the place_symbol
method is called and the active player's symbol is placed in the selected location.