Add check_stalemate method
This method uses a nested each/map on the board_array to create a new array containing the '#'
characters used to signify an empty space. I then use the .any? enumerable to check if there
are any '#" characters left in the array. If the array is full of player symbols, then the
game is declared a draw.