Refactor to fit Ruby style guide
I've created two classes, one for handling the board state
and the other to check for win conditions. I have split
all the class methodes to be complient with the style guide.
They're now short, simple, and acheive a single goal.

I have not yet finished the CheckWinner class. I also haven't
tested the functionality of the code in relation to the game
loop yet.