Fix GameBoard class methods
The @column variable was being referred to as @col in most places in the class.
I changed @column and all instances referring to it to @col.

The player_input method was expecting an array containing integers but was receiving
an array containing strings. I added a #map! to the passed array to convert the contents
to integers.

Finally, the display_board method wasn't outputting its contents in a way that worked
for the game flow. I realised I hadd used #index instead of #each on @board. The result
was display_board outputting a series of arrays rather than the contents. I have now used
nested #each methods and #print rather than #puts to achieve the desired result.