Fix player input bug
There was a bug in which the player would select a row and column but
during the output stage they would be reversed.

I fixed this by changing the order of @board[@col][@row] on both
place_symbol and move_validate to @board[@row][@col]. The way the array is constructed
means that each of the nested arrays represents a row and the
index of each element in each nested array correlates to the
column that element belongs to.